group catogory integration in taxonomy
